Use CoC for typescript stuff
This is a bit clunky, since ALE is also running and trying to do things.
But the CoC stuff for java/typescript is really powerful, so I'm using
that and just overriding some of the ALE shortcuts to use the CoC ones
when I'm in a js/ts file.

Mostly taken from this blog post:
https://medium.com/swlh/ultimate-vim-typescript-setup-35b5ac5c8c4e