Dotfiles @ 2021-12-10
Still WIP on the full CoC conversion for the vimrc, but in a pretty good
spot.  Use for c# and javascript/typescript is pretty good.  Still
haven't done much for python and firmware.

Biggest TODO is getting the clang-format on save back in firmware, since
that was an auto action from ALE